backend/pyproject.toml
[tool.poetry]
name = "user-mgt-app"
version = "0.1.0"
description = "User Mgt Application"
authors = ["karim baidar <karimbaidar@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
djangorestframework = "^3.12"
uwsgi = "^2.0"
mysqlclient = "^2.1"
django-cors-headers = "^4.3.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
pytest-django = "^4.2"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
